Fix error on Windows in heapq.pyx of issue #36 by specifying dtype
Addressing Issue #36 The expected type was 'Py_ssize_t', but 'long' was provided instead. This led to runtime exception. The solution was to explicitly set the dtype parameter to np.intp in the np.full function call. 

This modification resolves the error on Windows by ensuring the type matches the expected 'Py_ssize_t', as np.intp is the numpy data type that maps to 'ssize_t' on the platform being compiled on, thus aligning with Python's indexing types and C's ssize_t on Windows.

+++ pykonal/heapq.pyx
@@ -33,7 +33,7 @@ cdef class Heap(object):
     """
     def __init__(self, values):
         self.cy_values     = values
-        self.cy_heap_index = np.full(values.shape, fill_value=-1)
+        self.cy_heap_index = np.full(values.shape, fill_value=-1, dtype=np.intp)
 
 
     @property
